fix(expr): address all findings from quality evaluation
Implementation fixes:
- S1/PERF-1: Pre-parse format string expressions at FormatString::new() time
- S5: Empty lists default to list[nulltype] (compatible with any list type)
- S6: make_list validates max 2 nesting levels
- S8: Add RANGE_EXPR -> STRING and RANGE_EXPR -> LIST[INT] coercions
- S12: Rename ExprType::NULL to ExprType::NULLTYPE (matches OpenJD spec)
- BUG-2: Unify operation limit error type to OperationLimitExceeded
- BUG-3: URI path mapping case-insensitive scheme+authority per RFC 3986

Spec updates:
- S2: range-expr.md: descending ranges require negative step
- S3: function-library.md: fix call() signature
- S4: values.md: add cached size fields to enum definition
- S7: values.md: document hint_type and nested list promotion
- S9: function-library.md: fix EvalContext return types
- S10: function-library.md: document get_or_compile_regex
- S11: type-system.md: remove non-existent list constants
- S13: symbol-table.md: SymbolEntry -> SymbolTableEntry
- S14: values.md: Float64::from_str -> Float64::with_str
- S15: format-string.md: document resolve_typed, copy_used_symtab_values
- S16: evaluator.md: document per-evaluation regex cache

Test improvements:
- T1: Remove dead eval_fails helper
- T2: Remove 10 redundant is_list()-only tests, strengthen 2
- T3: Replace 7 p_err() checks with exact error message assertions
- T4: Replace 2 contains() assertions with exact values
- T5: Move 9 orphaned tests into #[cfg(test)] mod tests
- BUG-1/BUG-3: Add regression tests for bug fixes

Quality report updated with all findings resolved.

+++ crates/openjd-snapshots/src/hash_cache.rs
@@ -54,7 +54,7 @@ impl HashCache {
                     let hash: String = row.get(0)?;
                     let mtime: u64 = match row.get_ref(1)?.as_str() {
                         Ok(s) => s.parse::<u64>().unwrap_or(0),
-                        Err(_) => row.get::<_, u64>(1)?,
+                        Err(_) => row.get::<_, i64>(1)? as u64,
                     };
                     Ok((hash, mtime))
                 },
